requirements.txt
adbc_driver_snowflake
duckdb
pyarrow
python-dotenv
cryptography